Add configure-machine-for-blogging script
This script automates installation Jekyll on macOS and creates a blog subdirectory
in the directory that contains it. The installation follows the instructions listed
on <https://jekyllrb.com/docs/installation/macos/>.

It is safe to run this script more than once. That is, if Jekyll is already installed
it will not be re-installed and if there already exists a blog subdirectory then it
will not be overwritten.