Revert SHA ecb8f92459aa32fb4d20e827c2648bc3113754af
I decided to use a GitHub action to generate an publish the site into a dedicated branch, gh-pages,
instead of generating the site locally and then publishing in the master branch. As a result, the
site will be hosted from the root of the gh-pages branch. So, no /blog directory is needed.