Add commit CNAME step
Commit the CNAME file to the gh-pages branch so that GitHub knows to route dbates.com to the blog. Otherwise, GitHub will show a 404. Ideally this would be done in the same commit that pushes the built blog. For now, add a step to commit the CNAME file to the gh-pages branch.