Copy CNAME file to blog before building site
Add a new step to the github-pages action to copy the CNAME file to the blog
subdirectory and teach  Jekyll to copy that file into the built site. This
approach has the added benefit that the CNAME file and built site will be
committed in the same commit by the jekyll-action action and thus avoid a
possible service interruption that could occur between the time the site
is committed and when the CNAME file is committed.