Add script sitemap-changed to notify Google and Bing of a change to the sitemap.
The script takes an optional path to the previous sitemap to compare against the
current sitemap. Only if they diff will Google and Bing will be notified of the
new sitemap. Note that the current sitemap is assumed to be in the top-level directory
(which it is in the gh-pages branch).

I am not using this script right now, but I will be making use of it in a subsequent
change to the GitHub workflow action used to build the site.