Make my site more search engine friendly by reducing the length of the main page's <title> and providing custom descriptions.
According to Bing's webmaster tools, Bing does not want the title to exceed 70 characters. To achieve this,
specify a tagline in _config.yml. Otherwise the Jekyll SEO plugin builds the <title> for the home page by
concatenating the the title and description configuration properties. Additionally, Bing does not want the
<meta name="description"> to exceed 160 characters. By default Jekyll SEO generates the description from
the page content. These descriptions may not summarize the post well and may exceed Bing's character limit.
So, provide a custom description for each blog post to ensure each post has a good summary and doesn't
exceed the character limit.