Run sitemap-changed as part of the build workflow
Add a new step to the build workflow to determine if the sitemap has changed and submit the new sitemap to Google and Bing.
Google and Bing use the sitemap to know when to re-index the site.

After the GitHub Pages site is generated and pushed to the gh-pages branch I now
checkout the gh-pages branch into BuiltGitHubPages. (For now this checkout is done
inside the checkout of the main branch of the repository). Then sitemap-changed
is run.